Create new classes for CandidateDetail adapters instead of inner classes
